pyproject.toml
[project]
name = "lgtm-cli"
version = "0.1.0"
description = "Lightweight CLI for querying Loki, Prometheus/Mimir, and Tempo"
readme = "README.md"
authors = [
    { name = "Aiman Ismail", email = "aiman@primeintellect.ai" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "httpx>=0.28.1",
    "pyyaml>=6.0.3",
]

[project.scripts]
lgtm = "lgtm_cli.cli:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
